+++ game/shared/basecombatweapon_shared.h
@@ -88,7 +88,7 @@ class Color;
 
 namespace vgui2
 {
-	typedef unsigned long HFont;
+	typedef uint32 HFont;
 }
 
 // -----------------------------------------
